requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
requests==2.26.0
psycopg2-binary==2.9.1
Werkzeug==2.0.1
SQLAlchemy==1.4.46
pytest==7.4.0
